集数不能是float类型

+++ app/media/meta/metaanime.py
@@ -111,9 +111,9 @@ class MetaAnime(MetaBase):
                     end_episode = None
                 if begin_episode:
                     try:
-                        self.begin_episode = float(begin_episode)
-                        if end_episode and float(end_episode) != self.begin_episode:
-                            self.end_episode = float(end_episode)
+                        self.begin_episode = int(begin_episode)
+                        if end_episode and int(end_episode) != self.begin_episode:
+                            self.end_episode = int(end_episode)
                             self.total_episodes = (self.end_episode - self.begin_episode) + 1
                         else:
                             self.total_episodes = 1
